cheking

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -11,21 +11,21 @@
 int main(void)
 {
 int y;
-unsigned int u = 1;
-unsigned int p = 2;
-unsigned int k;
-printf("%u, ", u);
-printf("%u, ", p);
+long long int u = 1;
+long long int p = 2;
+long long int k;
+printf("%lld, ", u);
+printf("%lld, ", p);
 for (y = 1 ; y < 97 ; y++)
 {
 k = p + u;
 if (y != 96)
 {
-printf("%u, ", k);
+printf("%lld, ", k);
 }
 else
 {
-printf("%u", k);
+printf("%lld", k);
 }
 u = p;
 p = k;
